% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equiv.R
\name{atost}
\alias{atost}
\title{The alpha-TOST Corrective Procedure for (Bio)Equivalence Testing}
\usage{
atost(theta, sigma, nu, alpha, delta)
}
\arguments{
\item{theta}{A \code{numeric} value corresponding to the estimated parameter of interest (such as a difference of means).}

\item{sigma}{A \code{numeric} value corresponding to the estimated standard error of estimated \code{theta}.}

\item{nu}{A \code{numeric} value specifying the degrees of freedom.}

\item{alpha}{A \code{numeric} value specifying the significance level.}

\item{delta}{A \code{numeric} value or vector defining the (bio)equivalence margin(s). The procedure assumes symmetry, i.e., the (bio)equivalence region is \eqn{(-\delta, \delta)}.}
}
\value{
An object of class \code{tost} with the structure:
\itemize{
 \item \code{decision}:    A boolean variable indicating whether (bio)equivalence is accepted or not.
 \item \code{ci}:          Confidence region at the \eqn{1 - 2\alpha} level.
 \item \code{theta}:       The estimated difference(s) used in the test.
 \item \code{sigma}:       The estimated standard error used in the test.
 \item \code{nu}:          The number of degrees of freedom used in the test.
 \item \code{alpha}:       The significance level used in the test.
 \item \code{corrected_alpha}:       The significance level corrected by the adjustment.
 \item \code{delta}:       The (bio)equivalence limits used in the test.
 \item \code{method}:      The method used in the test (here the "alpha-TOST").
}
}
\description{
This function is used to compute the alpha-TOST, a corrective procedure of the significance level applied to the Two One-Sided Test (TOST) for (bio)equivalence testing in the univariate framework.
}
\examples{
data(skin)

theta_hat = diff(apply(skin,2,mean))
nu = nrow(skin) - 1
sig_hat = sd(apply(skin,1,diff))/sqrt(nu)
res_atost = cTOST:::atost(theta = theta_hat, sigma = sig_hat, nu = nu,
              alpha = 0.05, delta = log(1.25))
compare_to_tost(res_atost)
}
\author{
Younes Boulaguiem, Stéphane Guerrier, Dominique-Laurent Couturier
}
\keyword{internal}
