% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equiv.R
\name{BE}
\alias{BE}
\title{Assess Equivalence}
\usage{
BE(ci, c, ...)
}
\arguments{
\item{ci}{A \code{vector} with the lower and upper bounds of the confidence interval.}

\item{c}{A \code{numeric} value corresponding to equivalence limit. We assume symmetry, i.e, the equivalence interval corresponds to (-c,c)}

\item{...}{Additional parameters.}
}
\value{
The function returns a \code{numeric} value that corresponds to 0 is there is no equivalence, or 1 if there is equivalence.
}
\description{
Assess Equivalence
}
\author{
Younes Boulaguiem, Stéphane Guerrier, etc.
}
