% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equiv.R
\name{ci}
\alias{ci}
\title{Confidence Intervals in univariate TOST}
\usage{
ci(alpha, theta, sigma, nu, ...)
}
\arguments{
\item{alpha}{A \code{numeric} value specifying the significance level.}

\item{theta}{A \code{numeric} value corresponding to the estimated parameter of interest (such as a difference of means).}

\item{sigma}{A \code{numeric} value corresponding to the estimated standard error of estimated \code{theta}.}

\item{nu}{A \code{numeric} value specifying the degrees of freedom.}

\item{...}{Additional parameters.}
}
\value{
The function returns a numerical \code{vector} with the lower and upper bound of the confidence interval.
}
\description{
Confidence Intervals in univariate TOST
}
\author{
Younes Boulaguiem, Stéphane Guerrier, Dominique-Laurent Couturier
}
\keyword{internal}
