% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equiv.R
\name{coef.cTOST}
\alias{coef.cTOST}
\title{Extract relevant info to format}
\usage{
\method{coef}{cTOST}(x, ...)
}
\description{
Extract relevant info to format
}
\author{
Younes Boulaguiem, Stéphane Guerrier, etc.
}
