% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{compare_to_tost}
\alias{compare_to_tost}
\title{Comparison of a Corrective Procedure to the results of the Two One-Sided Tests (TOST) in Univariate Settings}
\usage{
compare_to_tost(x, ticks = 30, rn = 5)
}
\arguments{
\item{x}{A \code{tost} object, which is the output of one of the function: `ctost`.}

\item{ticks}{an integer indicating the number of segments that will be printed to represent the confidence intervals.}

\item{rn}{integer indicating the number of decimals places to be used (see function `round`) for the printed results.}
}
\value{
Prints a comparison between the TOST results (i.e., output of `tost`) and either the alpha-TOST or the delta-TOST results.
}
\description{
This function renders a comparison of the alpha-TOST or the delta-TOST outputs obtained with the function `ctost` to the TOST output obtained with `tost`.
}
\examples{
data(skin)
theta_hat = diff(apply(skin, 2, mean))
nu = nrow(skin) - 1
sig_hat = var(apply(skin, 1, diff)) / nu

# alpha-TOST
atost = ctost(theta = theta_hat, sigma = sig_hat, nu = nu,
              alpha = 0.05, delta = log(1.25), method = "alpha")
compare_to_tost(atost)

}
