% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equiv.R
\name{deltahat.fun}
\alias{deltahat.fun}
\title{Get Corrected (Bio)Equivalence Bounds}
\usage{
deltahat.fun(sigma, alpha, delta, nu)
}
\arguments{
\item{sigma}{A \code{numeric} value corresponding to the standard error.}

\item{alpha}{A \code{numeric} value specifying the significance level.}

\item{delta}{A \code{numeric} value corresponding to (bio)equivalence limit. We assume symmetry, i.e, the (bio)equivalence interval corresponds to (-delta,delta)}

\item{nu}{A \code{numeric} value corresponding to the number of degrees of freedom.}
}
\description{
This function applies the  delta-TOST corrective procedure to obtain the corrected (bio)equivalence bounds
}
\keyword{internal}
