% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_average_univariate.R
\name{dtost}
\alias{dtost}
\title{The delta-TOST Corrective Procedure for (Bio)Equivalence Testing}
\usage{
dtost(theta, sigma, nu, alpha, delta)
}
\arguments{
\item{theta}{A \code{numeric} value corresponding to the estimated parameter of interest (such as a difference of means).}

\item{sigma}{A \code{numeric} value corresponding to the estimated standard error of \code{theta}.}

\item{nu}{A \code{numeric} value specifying the degrees of freedom.}

\item{alpha}{A \code{numeric} value specifying the significance level (default: \code{alpha = 0.05}).}

\item{delta}{A \code{numeric} value defining the (bio)equivalence margin. The procedure assumes symmetry, i.e., the (bio)equivalence region is \eqn{(-\delta, \delta)}.}
}
\value{
An object of class \code{tost} with the following elements:
\itemize{
  \item \code{decision}: Logical; indicates whether (bio)equivalence is accepted.
  \item \code{ci}: Confidence region at the \eqn{1 - 2\alpha} level.
  \item \code{theta}: The estimated difference(s) used in the test.
  \item \code{sigma}: The estimated standard error used in the test.
  \item \code{nu}: The degrees of freedom used in the test.
  \item \code{alpha}: The significance level used in the test.
  \item \code{delta}: The (bio)equivalence limits used in the test.
  \item \code{corrected_delta}: The (bio)equivalence limits after adjustment.
  \item \code{method}: The method used in the test ("delta-TOST").
}
}
\description{
Computes the delta-TOST, a corrective procedure that adjusts the (bio)equivalence bounds applied to the Two One-Sided Test (TOST) for (bio)equivalence testing in the univariate framework.
}
\examples{
data(skin)

theta_hat = diff(apply(skin, 2, mean))
nu = nrow(skin) - 1
sig_hat = sd(apply(skin, 1, diff)) / sqrt(nu)
res_dtost = cTOST:::dtost(theta = theta_hat, sigma = sig_hat, nu = nu,
              alpha = 0.05, delta = log(1.25))
compare_to_tost(res_dtost)
}
\keyword{internal}
