% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equiv_multivariate.R
\name{find_sup_x}
\alias{find_sup_x}
\title{Size in multivariate xTOST}
\usage{
find_sup_x(alpha, Sigma, delta, seed = 10^5)
}
\arguments{
\item{alpha}{A \code{numeric} value specifying the significance level.}

\item{Sigma}{A \code{numeric} value (univariate) or \code{matrix} (multivariate) corresponding to the estimated variance of estimated \code{theta}.}

\item{delta}{A \code{numeric} value or vector defining the (bio)equivalence margin(s). The procedure assumes symmetry, i.e., the (bio)equivalence region is \eqn{(-\delta, \delta)}.}

\item{seed}{A \code{numeric} value specifying a seed for reproducibility (default: seed = \code{10^5}).}
}
\value{
The function returns a \code{numeric} vector that maximized the objective function.
}
\description{
This function is used to get the solution that maximizes the supremum for a constrained xTOST procedure.
}
\details{
The optimization method "Brent" is used for 2-dimensional or less problems only. The optimization method "Nelder-Mead" is used for higher dimensions.
}
\author{
Younes Boulaguiem, Luca Insolia, Stéphane Guerrier, Dominique-Laurent Couturier
}
\keyword{internal}
