% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equiv_quantile.R
\name{get_alpha_qTOST_MC}
\alias{get_alpha_qTOST_MC}
\title{Get alpha star for quantile TOST}
\usage{
get_alpha_qTOST_MC(
  gamma,
  n_x,
  n_y,
  pi_x,
  delta_l,
  delta_u,
  alpha,
  B,
  seed,
  tol = .Machine$double.eps,
  ...
)
}
\arguments{
\item{gamma}{A \code{numeric} value corresponding to the ratio of sample variance between the target and reference group.}

\item{n_x}{An \code{integer} value representing the sample size of the reference group.}

\item{n_y}{An \code{integer} value representing the sample size of the target group.}

\item{pi_x}{A \code{numeric} value defining the quantile of interest.}

\item{delta_l}{A \code{numeric} value defining the equivalence lower limit.}

\item{delta_u}{A \code{numeric} value defining the equivalence upper limit.}

\item{alpha}{A \code{numeric} value specifying the significance level.}

\item{B}{An \code{integer} value specifying the number of Monte Carlo replications.}

\item{seed}{An \code{integer} value specifying the random seed for reproducibility.}

\item{tol}{An optional \code{numeric} value defining the tolerance for root-finding (default: \code{tol = .Machine$double.eps^0.5}).}

\item{...}{Additional parameters}
}
\value{
A list with at least two components:
\itemize{
 \item \code{root}:       A \code{numeric} value corresponding to the location of the root.
 \item \code{f.root}:     A \code{numeric} value corresponding to the function evaluated at \code{root}.
}
}
\description{
Get alpha star for quantile TOST
}
\keyword{internal}
