% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_average_univariate.R
\name{get_alpha_star}
\alias{get_alpha_star}
\title{Compute Adjusted Significance Level (Alpha Star) for Univariate TOST}
\usage{
get_alpha_star(alpha = 0.05, sigma, nu, delta, ...)
}
\arguments{
\item{alpha}{A \code{numeric} value specifying the target significance level (default: \code{alpha = 0.05}).}

\item{sigma}{A \code{numeric} value representing the estimated standard error of \code{theta}.}

\item{nu}{A \code{numeric} value specifying the degrees of freedom.}

\item{delta}{A \code{numeric} value defining the (bio)equivalence margin. The procedure assumes symmetry, i.e., the (bio)equivalence region is \eqn{(-\delta, \delta)}.}

\item{...}{Additional parameters.}
}
\value{
A \code{numeric} value corresponding to the optimized (adjusted) significance level.
}
\description{
Calculates the adjusted significance level (\eqn{\alpha^*}) for the univariate Two One-Sided Tests (TOST) procedure.
}
\keyword{internal}
