% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equiv_multivariate.R
\name{get_alpha_TOST_MC_mv}
\alias{get_alpha_TOST_MC_mv}
\title{Get alpha star for multivariate TOST}
\usage{
get_alpha_TOST_MC_mv(
  alpha,
  Sigma,
  nu,
  delta,
  theta = NULL,
  B = 10^5,
  tol = .Machine$double.eps^0.5,
  seed = NULL,
  max_iter = 10,
  tolpower = NULL,
  ...
)
}
\arguments{
\item{alpha}{A \code{numeric} value specifying the significance level.}

\item{Sigma}{A \code{numeric} value (univariate) or \code{matrix} (multivariate) corresponding to the estimated variance of estimated \code{theta}.}

\item{nu}{A \code{numeric} value specifying the degrees of freedom. In the multivariate case, it is assumed to be the same across all dimensions.}

\item{delta}{A \code{numeric} value or vector defining the (bio)equivalence margin(s). The procedure assumes symmetry, i.e., the (bio)equivalence region is \eqn{(-\delta, \delta)}.}

\item{theta}{A \code{numeric} value specifying the initial value for the worst-case parameter configuration. If not provided, it's computed via find_sup_x().}

\item{B}{A \code{numeric} value specifying the number of Monte Carlo replication (default: B = \code{10^5}).}

\item{tol}{A \code{numeric} value specifying a tolerance level (default: tol = \code{.Machine$double.eps^0.5}).}

\item{seed}{A \code{character} value representing the use of default seed under \code{NULL}.}

\item{max_iter}{A \code{numeric} value specifying a maximum number of iteration.}

\item{tolpower}{A \code{numeric} value specifying the power level to optimize, see Details below for more information.}

\item{...}{Additional parameters.}
}
\description{
Get alpha star for multivariate TOST
}
\details{
Tolerance of power allowed between simulated power and nominal alpha. If not specified, it's automatically computed from the 1st and 99th percentiles of a binomial distribution.
}
\author{
Younes Boulaguiem, Luca Insolia, Stéphane Guerrier, Dominique-Laurent Couturier
}
\keyword{.}
\keyword{A}
\keyword{The}
\keyword{\code{alphas}:}
\keyword{\code{err_power}:}
\keyword{\code{iter}:}
\keyword{\code{list}}
\keyword{\code{min}:}
\keyword{\code{powers}:}
\keyword{\code{theta_sups}:}
\keyword{\code{theta_sup}:}
\keyword{\code{theta}}
\keyword{\eqn{\alpha}.}
\keyword{\item}
\keyword{\itemize{}
\keyword{a}
\keyword{absolute}
\keyword{and}
\keyword{between}
\keyword{corrected}
\keyword{corresponds}
\keyword{each}
\keyword{error}
\keyword{estimated}
\keyword{evaluated}
\keyword{evaluated.}
\keyword{final}
\keyword{function}
\keyword{in}
\keyword{internal}
\keyword{iteration.}
\keyword{iterations}
\keyword{level}
\keyword{level.}
\keyword{matrix}
\keyword{nominal}
\keyword{number}
\keyword{numerical}
\keyword{of}
\keyword{performed.}
\keyword{power}
\keyword{returns}
\keyword{significance}
\keyword{simulated}
\keyword{structure:}
\keyword{supremum.}
\keyword{that}
\keyword{the}
\keyword{to}
\keyword{used}
\keyword{value}
\keyword{values}
\keyword{variable}
\keyword{vector}
\keyword{with}
\keyword{}}
