% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equiv_multivariate.R
\name{get_alpha_TOST_MC_mv_core}
\alias{get_alpha_TOST_MC_mv_core}
\title{Get alpha star for multivariate TOST or xTOST}
\usage{
get_alpha_TOST_MC_mv_core(
  alpha,
  Sigma,
  nu,
  delta,
  theta = NULL,
  B = 10^5,
  tol = .Machine$double.eps^0.5,
  seed = NULL,
  argsup_meth = "x",
  ...
)
}
\arguments{
\item{alpha}{A \code{numeric} value specifying the significance level.}

\item{Sigma}{A \code{numeric} value (univariate) or \code{matrix} (multivariate) corresponding to the estimated variance of estimated \code{theta}.}

\item{nu}{A \code{numeric} value specifying the degrees of freedom. In the multivariate case, it is assumed to be the same across all dimensions.}

\item{delta}{A \code{numeric} value or vector defining the (bio)equivalence margin(s). The procedure assumes symmetry, i.e., the (bio)equivalence region is \eqn{(-\delta, \delta)}.}

\item{theta}{A \code{character} value specifying the value or vector representing the estimated difference(s) obtained by \code{argsup_meth} method under \code{NULL}.}

\item{B}{A \code{numeric} value specifying the number of Monte Carlo replication (default: B = \code{10^5}).}

\item{tol}{A \code{numeric} value specifying a tolerance level (default: tol = \code{.Machine$double.eps^0.5}).}

\item{seed}{A \code{character} value representing the use of default seed under \code{NULL}.}

\item{argsup_meth}{A \code{character} value representing the method to use (default: argsup_meth = \code{"x"}), see Details below for more information.}

\item{...}{Additional parameters.}
}
\value{
The function returns a \code{numeric} value that corresponds to the solution of the optimization.
}
\description{
Get alpha star for multivariate TOST or xTOST
}
\details{
The supremum of the size over the boundary of the equivalence region is found by direct optimisation (\code{argsup_meth = "x"}, the only implemented option).
The former is introduced in Boulaguiem et al. (2024, <https://doi.org/10.48550/arXiv.2411.16429>) and the latter is introduced in ...
}
\author{
Younes Boulaguiem, Luca Insolia, Stéphane Guerrier, Dominique-Laurent Couturier
}
\keyword{internal}
