% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{get_delta_TOST}
\alias{get_delta_TOST}
\title{Get Corrected (Bio)Equivalence Bounds}
\usage{
get_delta_TOST(
  alpha,
  sigma,
  nu,
  delta,
  l = 100,
  tol = .Machine$double.eps,
  ...
)
}
\arguments{
\item{alpha}{A \code{numeric} value specifying the significance level.}

\item{sigma}{A \code{numeric} value corresponding to the standard error.}

\item{nu}{A \code{numeric} value corresponding to the number of degrees of freedom.}

\item{delta}{A \code{numeric} value corresponding to (bio)equivalence limit. We assume symmetry, i.e, the (bio)equivalence interval corresponds to (-delta,delta)}
}
\description{
This function applies the  delta-TOST corrective procedure to obtain the corrected (bio)equivalence bounds
}
\keyword{internal}
