% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equiv_multivariate.R
\name{get_interval}
\alias{get_interval}
\title{Confidence interval search in multivariate xTOST}
\usage{
get_interval(delta, Sigma, cte, alpha, interval, B)
}
\arguments{
\item{delta}{A \code{numeric} value or vector defining the (bio)equivalence margin(s). The procedure assumes symmetry, i.e., the (bio)equivalence region is \eqn{(-\delta, \delta)}.}

\item{Sigma}{A \code{numeric} value (univariate) or \code{matrix} (multivariate) corresponding to the estimated variance of estimated \code{theta}.}

\item{cte}{A \code{numeric} value specifying ... (the constraint?).}

\item{alpha}{A \code{numeric} value specifying the significance level.}

\item{interval}{A \code{numeric} vector specifying range of interest to search.}

\item{B}{A \code{numeric} value specifying the number of point to search.}
}
\value{
The function returns a \code{numeric} vector that minimized the objective function.
}
\description{
This function performs a grid search over a specified interval to
find the interval where the size is maximized.
}
\author{
Younes Boulaguiem, Luca Insolia, Stéphane Guerrier, Dominique-Laurent Couturier
}
\keyword{internal}
