% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equiv_quantile.R
\name{get_qTOST_rvs}
\alias{get_qTOST_rvs}
\title{Generate distribution of estimated parameter of interest}
\usage{
get_qTOST_rvs(theta, gamma, n_x, n_y, pi_x, B = 10^4, seed = 12345)
}
\arguments{
\item{theta}{A \code{numeric} value corresponding to the parameter of interest.}

\item{gamma}{A \code{numeric} value corresponding to the ratio of sample variance between the target and reference group.}

\item{n_x}{An \code{integer} value representing the sample size of the reference group.}

\item{n_y}{An \code{integer} value representing the sample size of the target group.}

\item{pi_x}{A \code{numeric} value defining the quantile of interest.}

\item{B}{An optional \code{integer} value specifying the number of Monte Carlo replications (default: \code{B = 10^5}).}

\item{seed}{An optional \code{integer} value specifying the random seed for reproducibility (default: \code{seed = 12345}).}
}
\value{
A list with the structure:
\itemize{
 \item theta_hat:         B \code{numeric} values corresponding to the \eqn{\hat{\theta}} used in the test.
 \item sigma_theta_hat:   B \code{numeric} values corresponding to the \eqn{\hat{\sigma}} of \eqn{\hat{\theta}} used in the test.
}
}
\description{
This function is used to generate estimated parameter of interest and its standard deviation
using Monte Carlo simulation.
}
\keyword{internal}
