% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{Generate ggplot2-like Colors}
\usage{
gg_color_hue(n)
}
\arguments{
\item{n}{The number of colors to generate.}
}
\value{
A character vector of hex color codes.
}
\description{
Creates a vector of colors matching the default ggplot2 palette
by spacing hues evenly around the HCL color wheel.
}
\keyword{internal}
