% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{\%||\%}
\alias{\%||\%}
\title{Null-Coalescing Operator}
\usage{
a \%||\% b
}
\arguments{
\item{a}{The value to check for NULL.}

\item{b}{The default value to return if 'a' is NULL.}
}
\value{
The first non-NULL value from left to right.
}
\description{
Returns the left-hand side if it is not NULL, otherwise returns
the right-hand side. This provides a convenient shorthand for setting default values.
}
\keyword{internal}
