% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_average_univariate.R
\name{obj_fun_alpha_star}
\alias{obj_fun_alpha_star}
\title{Objective Function for Optimization in Univariate TOST}
\usage{
obj_fun_alpha_star(test, alpha = 0.05, sigma, nu, delta, ...)
}
\arguments{
\item{test}{A \code{numeric} value specifying the significance level to be optimized.}

\item{alpha}{A \code{numeric} value specifying the target significance level (default: \code{alpha = 0.05}).}

\item{sigma}{A \code{numeric} value representing the estimated standard error of \code{theta}.}

\item{nu}{A \code{numeric} value specifying the degrees of freedom.}

\item{delta}{A \code{numeric} value defining the (bio)equivalence margin. The procedure assumes symmetry, i.e., the (bio)equivalence region is \eqn{(-\delta, \delta)}.}

\item{...}{Additional parameters.}
}
\value{
A \code{numeric} value representing the objective function to be minimized.
}
\description{
Computes the objective function used to optimize the significance level in the univariate Two One-Sided Tests (TOST) procedure.
}
\keyword{internal}
