% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_average_univariate.R
\name{obj_fun_alpha_TOST}
\alias{obj_fun_alpha_TOST}
\title{Objective Function of the alpha-TOST Corrective Procedure}
\usage{
obj_fun_alpha_TOST(test, alpha, sigma, nu, delta, theta = NULL, ...)
}
\arguments{
\item{test}{A \code{numeric} value specifying the significance level to optimize.}

\item{alpha}{A \code{numeric} value specifying the significance level.}

\item{sigma}{A \code{numeric} value corresponding to the estimated standard error of estimated \code{theta}.}

\item{nu}{A \code{numeric} value corresponding to the number of degrees of freedom.}

\item{delta}{A \code{numeric} value or vector defining the (bio)equivalence margin(s). The procedure assumes symmetry, i.e., the (bio)equivalence region is \eqn{(-\delta, \delta)}.}

\item{theta}{A \code{numeric} value representing the estimated difference(s) (e.g., between a generic and reference product) or a \code{character} value representing the use of equivalence margin(s) for \eqn{\theta} under \code{NULL} (e.g., \eqn{\theta} = \eqn{\delta}).}

\item{...}{Additional parameters.}
}
\value{
The function returns a \code{numeric} value for the objective function.
}
\description{
Objective Function of the alpha-TOST Corrective Procedure
}
\keyword{internal}
