% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equiv_multivariate.R
\name{obj_fun_alpha_TOST_MC_mv}
\alias{obj_fun_alpha_TOST_MC_mv}
\title{Objective function to optimize for multivariate TOST}
\usage{
obj_fun_alpha_TOST_MC_mv(
  test,
  alpha,
  Sigma,
  nu,
  delta,
  theta = NULL,
  B = 10^5,
  seed = NULL,
  ...
)
}
\arguments{
\item{test}{A \code{numeric} value specifying the significance level to optimize.}

\item{alpha}{A \code{numeric} value specifying the significance level.}

\item{Sigma}{A \code{numeric} value (univariate) or \code{matrix} (multivariate) corresponding to the estimated variance of estimated \code{theta}.}

\item{nu}{A \code{numeric} value specifying the degrees of freedom. In the multivariate case, it is assumed to be the same across all dimensions.}

\item{delta}{A \code{numeric} value or vector defining the (bio)equivalence margin(s). The procedure assumes symmetry, i.e., the (bio)equivalence region is \eqn{(-\delta, \delta)}.}

\item{theta}{A \code{numeric} value representing the estimated difference(s) (e.g., between a generic and reference product) or a \code{character} value representing the use of equivalence margin(s) for \eqn{\theta} under \code{NULL} (e.g., \eqn{\theta} = \eqn{(-\delta, \delta)}).}

\item{B}{A \code{numeric} value specifying the number of Monte Carlo replication (default: B = \code{10^5}).}

\item{seed}{A \code{numeric} value specifying a seed for reproducibility representing the use of multivariate power or a \code{character} value representing the use of univariate power under \code{NULL}.}

\item{...}{Additional parameters.}
}
\value{
The function returns a \code{numeric} value for the objective function.
}
\description{
Objective function to optimize for multivariate TOST
}
\author{
Younes Boulaguiem, Luca Insolia, Stéphane Guerrier, Dominique-Laurent Couturier
}
\keyword{internal}
