% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equiv.R
\name{obj_fun_delta_hat}
\alias{obj_fun_delta_hat}
\title{Objective Function of the delta-TOST Corrective Procedure}
\usage{
obj_fun_delta_hat(delta_star, sigma, alpha, delta, nu)
}
\arguments{
\item{delta_star}{The (bio)equivalence bound parameter to be optimized.}

\item{sigma}{The considered standard error.}

\item{alpha}{The nominal level for the test.}

\item{delta}{The (bio)equivalence bound used for the TOST decision.}

\item{nu}{The degrees of freedom parameter.}
}
\description{
Objective Function of the delta-TOST Corrective Procedure
}
\keyword{internal}
