% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equiv_multivariate.R
\name{obj_fun_xTOST_constr_size_mv}
\alias{obj_fun_xTOST_constr_size_mv}
\title{Objective function to optimize for multivariate xTOST}
\usage{
obj_fun_xTOST_constr_size_mv(test, Sigma, cte, delta, alpha)
}
\arguments{
\item{test}{A \code{numeric} value specifying the significance level to optimize.}

\item{Sigma}{A \code{numeric} value (univariate) or \code{matrix} (multivariate) corresponding to the estimated variance of estimated \code{theta}.}

\item{cte}{A \code{numeric} value specifying ...(? the constraint?)}

\item{delta}{A \code{numeric} value or vector defining the (bio)equivalence margin(s). The procedure assumes symmetry, i.e., the (bio)equivalence region is \eqn{(-\delta, \delta)}.}

\item{alpha}{A \code{numeric} value specifying the significance level.}
}
\value{
The function returns a \code{numeric} value for the objective function.
}
\description{
Objective function to optimize for multivariate xTOST
}
\author{
Younes Boulaguiem, Luca Insolia, Stéphane Guerrier, Dominique-Laurent Couturier
}
\keyword{internal}
