% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equiv_quantile.R
\name{power_qTOST_MC}
\alias{power_qTOST_MC}
\title{Power calculation for quantile TOST using Monte Carlo simulation}
\usage{
power_qTOST_MC(
  theta,
  gamma = gamma,
  n_x,
  n_y,
  pi_x,
  delta_l,
  delta_u,
  alpha,
  B,
  seed,
  ...
)
}
\arguments{
\item{theta}{A \code{numeric} value corresponding to the parameter of interest.}

\item{gamma}{A \code{numeric} value corresponding to the ratio of sample variance between the target and reference group.}

\item{n_x}{An \code{integer} value representing the sample size of the reference group.}

\item{n_y}{An \code{integer} value representing the sample size of the target group.}

\item{pi_x}{A \code{numeric} value defining the quantile of interest.}

\item{delta_l}{A \code{numeric} value defining the equivalence lower limit.}

\item{delta_u}{A \code{numeric} value defining the equivalence upper limit.}

\item{alpha}{A \code{numeric} value specifying the significance level.}

\item{B}{An \code{integer} value specifying the number of Monte Carlo replications.}

\item{seed}{An \code{integer} value specifying the random seed for reproducibility.}

\item{...}{Additional parameters}
}
\value{
The function returns a \code{numeric} value that corresponds to a probability.
}
\description{
This function estimates the statistical power of the quantile-based
Two One-Sided Tests (qTOST) procedure using Monte Carlo simulations.
}
\details{
The estimated power is the proportion of Monte Carlo replicates in which equivalence is concluded.
}
\keyword{internal}
