% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_average_univariate.R
\name{power_TOST}
\alias{power_TOST}
\title{Power Function of Univariate TOST}
\usage{
power_TOST(alpha, theta, sigma, nu, delta, ...)
}
\arguments{
\item{alpha}{A \code{numeric} value specifying the significance level.}

\item{theta}{A \code{numeric} value representing the parameter of interest (e.g., a difference of means).}

\item{sigma}{A \code{numeric} value representing the estimated standard error of \code{theta}.}

\item{nu}{A \code{numeric} value specifying the degrees of freedom.}

\item{delta}{A \code{numeric} value defining the (bio)equivalence margin. The procedure assumes symmetry, i.e., the (bio)equivalence region is \eqn{(-\delta, \delta)}.}

\item{...}{Additional parameters.}
}
\value{
A \code{numeric} value corresponding to the probability (power) of the TOST procedure.
}
\description{
Computes the power function for the univariate Two One-Sided Tests (TOST) procedure.
}
\keyword{internal}
