% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{power_xTOST}
\alias{power_xTOST}
\title{TITLE}
\usage{
power_xTOST(theta, sig_hat, delta, ...)
}
\arguments{
\item{theta}{...}

\item{sig_hat}{...}

\item{delta}{...}

\item{...}{Additional parameters.}
}
\description{
What this function does
}
\keyword{internal}
