% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equiv_multivariate.R
\name{power_xTOST_mv}
\alias{power_xTOST_mv}
\title{Power function for multivariate xTOST procedure}
\usage{
power_xTOST_mv(theta, delta, Sigma, alpha = 1/2, seed = 10^5)
}
\arguments{
\item{theta}{A \code{numeric} value or vector representing the estimated difference(s) (e.g., between a generic and reference product).}

\item{delta}{A \code{numeric} value or vector defining the (bio)equivalence margin(s). The procedure assumes symmetry, i.e., the (bio)equivalence region is \eqn{(-\delta, \delta)}. In the multivariate case, it is assumed to be the same across all dimensions.}

\item{Sigma}{A \code{numeric} value (univariate) or \code{matrix} (multivariate) corresponding to the estimated variance of estimated \code{theta}.}

\item{alpha}{A \code{numeric} value specifying the significance level (default: alpha = \code{0.5}).}

\item{seed}{A \code{numeric} value specifying a seed for reproducibility (default: seed = \code{10^5}).}
}
\value{
The function returns a \code{numeric} value that corresponds to a probability.
}
\description{
This function is used to approximate the power by a standardized multivariate normal vector lies within the (bio)equivalence margins.
}
\author{
Younes Boulaguiem, Luca Insolia, Stéphane Guerrier, Dominique-Laurent Couturier
}
\keyword{internal}
