% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equiv.R
\name{print.cTOST}
\alias{print.cTOST}
\title{Print the results of the equivalence assessment}
\usage{
\method{print}{cTOST}(x, ...)
}
\description{
Print the results of the equivalence assessment
}
\author{
Younes Boulaguiem, Stéphane Guerrier, etc.
}
