% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_mean_test.R
\name{print.dp_tost_mean}
\alias{print.dp_tost_mean}
\title{Print method for DP-TOST mean test}
\usage{
\method{print}{dp_tost_mean}(x, ticks = 30, rn = 5, ...)
}
\arguments{
\item{x}{An object of class \code{dp_tost_mean}}

\item{ticks}{Number of characters for visual representation}

\item{rn}{Number of decimal places for rounding}

\item{...}{Additional arguments (not used)}
}
\value{
Invisibly returns the input object
}
\description{
Print method for DP-TOST mean test
}
