% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_proportion_test.R
\name{print.dp_tost_prop}
\alias{print.dp_tost_prop}
\title{Print Results of DP-TOST for Proportions}
\usage{
\method{print}{dp_tost_prop}(x, ticks = 30, rn = 5, ...)
}
\arguments{
\item{x}{A \code{dp_tost_prop} object returned by \code{prop_test_equiv_dp}}

\item{ticks}{Number of ticks to print the confidence interval in the console. Default is 30.}

\item{rn}{Number of digits to consider when printing the results. Default is 5.}

\item{...}{Further arguments to be passed to or from methods.}
}
\value{
Prints object invisibly.
}
\description{
Print Results of DP-TOST for Proportions
}
