% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_qnt.R
\name{print.m_qtost}
\alias{print.m_qtost}
\title{Print Results of (Bio)Equivalence testing in Two Quantiles Setting.}
\usage{
\method{print}{m_qtost}(x, ticks = 60, rn = 5, ...)
}
\arguments{
\item{x}{A \code{qtost} object, which is the output of the function 'qtost'.}

\item{ticks}{Number of ticks to print the confidence interval in the console.}

\item{rn}{Number of digits to consider when printing the results.}

\item{...}{Further arguments to be passed to or from methods.}
}
\value{
Prints object.
}
\description{
Print Results of (Bio)Equivalence testing in Two Quantiles Setting.
}
