% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_mean_test.R
\name{print.tost_dp}
\alias{print.tost_dp}
\title{Print method for DP-TOST mean test}
\usage{
\method{print}{tost_dp}(x, ticks = 30, rn = 5, ...)
}
\arguments{
\item{x}{An object of class \code{tost_dp}}

\item{ticks}{Number of characters for visual representation}

\item{rn}{Number of decimal places for rounding}

\item{...}{Additional arguments (not used)}
}
\value{
Invisibly returns the input object
}
\description{
Print method for DP-TOST mean test
}
