% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{skin}
\alias{skin}
\title{Log-transformed cutaneous delivery of econazole (ECZ) from bioequivalent products on porcine skin}
\format{
A `data.frame` with 17 rows and 2 columns:
\describe{
  \item{Reference}{Log-transformed econazole nitrate delivery for the reference product.}
  \item{Generic}{Log-transformed econazole nitrate delivery for the generic bioequivalent product.}
}
}
\usage{
data(skin)
}
\description{
Original data were collected using the cutaneous biodistribution method described in Quartier et. al. (2019),
and represents cutaneous delivery of econazole nitrate (ECZ in ng/cm^2) on porcine skin from a reference medicinal
product and an approved bioequivalent product. The dataset contains 17 pairs of comparable porcine skin samples on which
measurement of ECZ deposition was gathered, and log transformed, using both creams.
}
\examples{
data(skin)
theta <- diff(apply(skin,2,mean))
n <- nrow(skin)
nu <- n - 1
sigma <- var(apply(skin,1,diff)) / n
}
\references{
Quartier J. et al. "Cutaneous Biodistribution: A High-Resolution Methodology
to Assess Bioequivalence in Topical Skin Delivery", Pharmaceutics, 2019.

Boulaguiem Y. et al. "Finite Sample Adjustments for Average Equivalence Testing", Statistics in Medicine, 2024.
}
\keyword{datasets}
