% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_average_univariate.R
\name{ci}
\alias{ci}
\title{Confidence Interval for Univariate TOST}
\usage{
ci(alpha, theta, sigma, nu, ...)
}
\arguments{
\item{alpha}{A \code{numeric} value specifying the significance level.}

\item{theta}{A \code{numeric} value representing the estimated parameter of interest (e.g., a difference of means).}

\item{sigma}{A \code{numeric} value representing the estimated standard error of \code{theta}.}

\item{nu}{A \code{numeric} value specifying the degrees of freedom.}

\item{...}{Additional parameters.}
}
\value{
A numeric \code{vector} containing the lower and upper bounds of the confidence interval.
}
\description{
Computes the confidence interval for the univariate Two One-Sided Tests (TOST) procedure.
}
\keyword{internal}
