% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equiv.R
\name{ci}
\alias{ci}
\title{Confidence Intervals}
\usage{
ci(alpha, theta, sigma_nu, nu, ...)
}
\arguments{
\item{alpha}{A \code{numeric} value specifying the significance level (default = \code{0.05}).}

\item{theta}{A \code{numeric} value corresponding to the estimated parameter of interest (such as a difference of means).}

\item{sigma_nu}{A \code{numeric} value corresponding to the estimated standard error.}

\item{nu}{A \code{numeric} value corresponding to the number of degrees of freedom.}

\item{...}{Additional parameters.}
}
\value{
The function returns a numerical \code{vector} with the lower and upper bound of the confidence intervals.
}
\description{
Confidence Intervals
}
\author{
Younes Boulaguiem, Stéphane Guerrier, Dominique-Laurent Couturier
}
\keyword{internal}
