% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_qnt.R
\name{compare_to_qtost}
\alias{compare_to_qtost}
\title{Comparison of a Corrective Procedure to the Results of the Quantile Two One-Sided Tests (qTOST) in Single Quantile Setting}
\usage{
compare_to_qtost(...)
}
\arguments{
\item{x}{A \code{qtost} object, which is the output of one of the function: `qtost`.}

\item{ticks}{an integer indicating the number of segments that will be printed to represent the confidence intervals.}

\item{rn}{integer indicating the number of decimals places to be used (see function `round`) for the printed results.}
}
\value{
Pints a comparison between the qTOST results (i.e., output of `qtost`) and the alpha-qTOST results.
}
\description{
This function renders a comparison of the qTOST or the alpha-qTOST outputs obtained with the function `qtost`.
}
