% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equiv.R
\name{ctost}
\alias{ctost}
\title{Finite Sample Adjusted (Bio)Equivalence Testing}
\usage{
ctost(
  theta,
  sigma,
  nu,
  delta,
  alpha = 0.05,
  method,
  B = 10^4,
  seed = 101010,
  correction = "offline",
  ...
)
}
\arguments{
\item{theta}{A \code{numeric} value corresponding to the difference of means.}

\item{sigma}{A \code{numeric} value corresponding to the standard error in univariate setting, the estimated covariance matrix in multivariate setting.}

\item{nu}{A \code{numeric} value corresponding to the number of degrees of freedom.}

\item{delta}{A \code{numeric} value corresponding to (bio)equivalence limit. We assume symmetry, i.e, the (bio)equivalence interval corresponds to (-delta,delta).}

\item{alpha}{A \code{numeric} value specifying the significance level (default: alpha = \code{0.05}).}

\item{method}{A \code{character} value corresponding to the considered finite sample adjustment method, see Details below for more information.}

\item{B}{A \code{numeric} value specifying the number of Monte Carlo replication (default: B = \code{10^4}).}

\item{seed}{A \code{numeric} value specifying a seed for reproducibility (default: seed = \code{101010}).}

\item{correction}{A \code{character} value corresponding to the considered correction method, see Details below for more information (default: correction = \code{"offline"}).}

\item{...}{Additional parameters.}
}
\value{
A \code{tost} object with the structure:
\itemize{
 \item decision:    A boolean variable indicating whether (bio)equivalence is accepted or not.
 \item CI:          Confidence interval at the 1 - 2*alpha level.
 \item theta:       The difference of means used for the test.
 \item sigma:       The standard error used for the test.
 \item nu:          The number of degrees of freedom used for the test.
 \item alpha:       The significance level used for the test.
 \item delta:       The (bio)equivalence limits used for the test.
 \item method:      The method used for the test (optimal, alpha-TOST and delta-TOST).
 \item setting:     The setting used (univariate or multivariate).
}
}
\description{
This function is used to compute finite sample corrected version of the standard (univariate or multivariate) TOST.
}
\examples{
data(skin)

theta_hat = diff(apply(skin,2,mean))
nu = nrow(skin) - 1
sig_hat = var(apply(skin,1,diff))/nu

# alpha-TOST
atost = ctost(theta = theta_hat, sigma = sig_hat, nu = nu,
              alpha = 0.05, delta = log(1.25), method = "alpha")
atost
compare_to_tost(atost)

# delta-TOST
dtost = ctost(theta = theta_hat, sigma = sig_hat, nu = nu,
              alpha = 0.05, delta = log(1.25), method = "delta")
dtost
compare_to_tost(dtost)
}
