% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equiv.R
\name{ctost}
\alias{ctost}
\title{Finite Sample Adjusted (Bio)Equivalence Testing}
\usage{
ctost(theta, sigma, nu, delta, alpha = 0.05, method = "alpha", B = 10^4, ...)
}
\arguments{
\item{theta}{A \code{numeric} value corresponding to the difference of means.}

\item{sigma}{A \code{numeric} value corresponding to the standard error.}

\item{nu}{A \code{numeric} value corresponding to the number of degrees of freedom.}

\item{delta}{A \code{numeric} value corresponding to (bio)equivalence limit. We assume symmetry, i.e, the (bio)equivalence interval corresponds to (-delta,delta)}

\item{alpha}{A \code{numeric} value specifying the significance level (default: alpha = 0.05).}

\item{method}{A \code{character} value corresponding to the considered finite sample adjustment method, see Details below for more information.}

\item{B}{TO DOCUMENT ???}

\item{...}{Additional parameters.}
}
\value{
A \code{tost} object with the structure:
\itemize{
 \item decision:    A boolean variable indicating whether (bio)equivalence is accepted or not.
 \item ci:          Confidence interval at the 1 - 2*alpha level.
 \item theta:       The difference of means used for the test.
 \item sigma:       The standard error used for the test.
 \item nu:          The number of degrees of freedom used for the test.
 \item alpha:       The significance level used for the test.
 \item delta:       The (bio)equivalence limits used for the test.
 \item method:      The method used for the test (alpha-TOST and delta-TOST).
 \item setting:     The setting used (univariate or multivariate)
}
}
\description{
This function is used to compute finite sample corrected version of the standard (univariate or multivariate) TOST.
}
\details{
In univariate settings, two methods are available: alpha-TOST (using method = 'alpha') and delta-TOST (using method = 'delta').
The alpha-TOST and delta-TOST are introduced in Boulaguiem et al. (2024, <https://doi.org/10.1002/sim.9993>). The former is a corrective procedure of the significance level applied to the TOST while the latter
adjusts the equivalence limits. In general, the alpha-TOST appears to outperform the delta-TOST.

In multivariate setting, the only available method is the (multivariate) alpha-TOST (using method = 'alpha') introduced in Boulaguiem et al. (2024, bioarxiv).
}
\examples{
data(skin)

theta_hat = diff(apply(skin,2,mean))
nu = nrow(skin) - 1
sig_hat = sd(apply(skin,1,diff))/sqrt(nu)

# alpha-TOST
atost = ctost(theta = theta_hat, sigma = sig_hat, nu = nu,
              alpha = 0.05, delta = log(1.25), method = "alpha")
atost
compare_to_tost(atost)

# delta-TOST
dtost = ctost(theta = theta_hat, sigma = sig_hat, nu = nu,
              alpha = 0.05, delta = log(1.25), method = "delta")
dtost
compare_to_tost(dtost)
}
\author{
Younes Boulaguiem, Stéphane Guerrier, Dominique-Laurent Couturier, Luca Insolia
}
