% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equiv.R
\name{ctost}
\alias{ctost}
\title{Finite Sample Adjusted (Bio)Equivalence Testing}
\usage{
ctost(
  theta,
  sigma,
  nu,
  delta,
  alpha = 0.05,
  method,
  B = 10^4,
  seed = 101010,
  correction = "offline",
  ...
)
}
\arguments{
\item{theta}{A \code{numeric} value or vector representing the estimated difference(s) (e.g., between a generic and reference product).}

\item{sigma}{A \code{numeric} value (univariate) or \code{matrix} (multivariate) corresponding to the estimated variance of estimated \code{theta}.}

\item{nu}{A \code{numeric} value specifying the degrees of freedom. In the multivariate case, it is assumed to be the same across all dimensions.}

\item{delta}{A \code{numeric} value or vector defining the (bio)equivalence margin(s). The procedure assumes symmetry, i.e., the (bio)equivalence region is \eqn{(-\delta, \delta)}. In the multivariate case, it is assumed to be the same across all dimensions.}

\item{alpha}{A \code{numeric} value specifying the significance level (default: alpha = \code{0.05}).}

\item{method}{A \code{character} value corresponding to the considered finite sample adjustment method, see Details below for more information.}

\item{B}{A \code{numeric} value specifying the number of Monte Carlo replication (default: B = \code{10^4}).}

\item{seed}{A \code{numeric} value specifying a seed for reproducibility (default: seed = \code{101010}).}

\item{correction}{A \code{character} value corresponding to the considered correction method, see Details below for more information (default: correction = \code{"offline"}).}

\item{...}{Additional parameters.}
}
\value{
A \code{ctost} object with the structure:
\itemize{
 \item \code{decision}:    A boolean variable indicating whether (bio)equivalence is accepted or not.
 \item \code{ci}:          Confidence region at the \eqn{1 - 2\alpha} level.
 \item \code{theta}:       The estimated difference(s) used in the test.
 \item \code{sigma}:       The estimated variance of \code{theta}, a \code{numeric} in univariate settings or \code{matrix} in multivariate settings.
 \item \code{nu}:          The number of degrees of freedom used in the test.
 \item \code{alpha}:       The significance level used in the test.
 \item \code{corrected_alpha}: The significance level corrected by the adjustment ("alpha-TOST" used).
 \item \code{corrected_delta}: The (bio)equivalence limits corrected by the adjustment ("delta-TOST" used).
 \item \code{delta}:       The (bio)equivalence limits used in the test.
 \item \code{method}:      The method used in the test (optimal, alpha-TOST and delta-TOST).
 \item \code{setting}:     The setting used (univariate or multivariate).
}
}
\description{
This function is used to compute finite sample corrected version of the standard (univariate or multivariate) TOST.
}
\examples{
data(skin)

theta_hat = diff(apply(skin,2,mean))
nu = nrow(skin) - 1
sig_hat = var(apply(skin,1,diff))/nu

# alpha-TOST
atost = ctost(theta = theta_hat, sigma = sig_hat, nu = nu,
              alpha = 0.05, delta = log(1.25), method = "alpha")
atost
compare_to_tost(atost)

# delta-TOST
dtost = ctost(theta = theta_hat, sigma = sig_hat, nu = nu,
              alpha = 0.05, delta = log(1.25), method = "delta")
dtost
compare_to_tost(dtost)
}
\author{
Younes Boulaguiem, Stéphane Guerrier, Dominique-Laurent Couturier
}
