% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equiv.R
\name{cTOST}
\alias{cTOST}
\title{Equivalence Assessment in the univariate framework for the TOST and the alpha-TOST}
\usage{
cTOST(alpha, theta, sigma_nu, nu, delta, ...)
}
\arguments{
\item{alpha}{A \code{numeric} value specifying the significance level (default = \code{0.05}).}

\item{theta}{A \code{numeric} value corresponding to the estimated parameter of interest (such as a difference of means).}

\item{sigma_nu}{A \code{numeric} value corresponding to the the estimated standard error.}

\item{nu}{A \code{numeric} value corresponding to the number of degrees of freedom.}

\item{delta}{A \code{numeric} value corresponding to equivalence limit. We assume symmetry, i.e, the equivalence interval corresponds to (-delta,delta)}

\item{...}{Additional parameters.}
}
\value{
The function returns an object of class `matrix`.
}
\description{
The function returns many useful components in the assessment of bioequivalence
in the univariate frameowrk for each of the TOST and alphaTOST, namely, the estimated parameter,
the estimated standard error, confidence intervals for both methods, the equivalence limit,
the original significance level and the corrected one.
}
\examples{
theta <- diff(apply(skin,2,mean))
nu <- nrow(skin)-1
sigma_nu <- sd(apply(skin,1,diff))/sqrt(nu)
alpha <- 0.05
delta <- log(1.25)
cTOST(alpha, theta, sigma_nu, nu, delta)
}
\author{
Younes Boulaguiem, Stéphane Guerrier, etc.
}
