% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equiv.R
\name{dtost}
\alias{dtost}
\title{The delta-TOST Corrective Procedure for (Bio)Equivalence Testing}
\usage{
dtost(theta, sigma, nu, alpha, delta)
}
\arguments{
\item{theta}{A \code{numeric} value corresponding to the estimated parameter of interest (such as a difference of means).}

\item{sigma}{A \code{numeric} value corresponding to the estimated standard error of estimated \code{theta}.}

\item{nu}{A \code{numeric} value specifying the degrees of freedom.}

\item{alpha}{A \code{numeric} value specifying the significance level (default: alpha = \code{0.05}).}

\item{delta}{A \code{numeric} value or vector defining the (bio)equivalence margin(s). The procedure assumes symmetry, i.e., the (bio)equivalence region is \eqn{(-\delta, \delta)}.}
}
\value{
A \code{dtost} object with the structure:
\itemize{
 \item \code{decision}:    A boolean variable indicating whether (bio)equivalence is accepted or not.
 \item \code{ci}:          Confidence region at the \eqn{1 - 2\alpha} level.
 \item \code{theta}:       The estimated difference(s) used in the test.
 \item \code{sigma}:       The estimated standard error used in the test.
 \item \code{nu}:          The number of degrees of freedom used in the test.
 \item \code{alpha}:       The significance level used in the test.
 \item \code{delta}:       The (bio)equivalence limits used in the test.
 \item \code{corrected_delta}: The (bio)equivalence limits corrected by the adjustment.
 \item \code{method}:      The method used in the test (here the "delta-TOST").
}
}
\description{
This function is used to compute the delta-TOST, a corrective procedure of the (bio)equivalence bounds applied to the Two One-Sided Test (TOST) for (bio)equivalence testing in the univariate framework.
}
\examples{
data(skin)

theta_hat = diff(apply(skin,2,mean))
nu = nrow(skin) - 1
sig_hat = sd(apply(skin,1,diff))/sqrt(nu)
res_dtost = cTOST:::dtost(theta = theta_hat, sigma = sig_hat, nu = nu,
              alpha = 0.05, delta = log(1.25))
compare_to_tost(res_dtost)

}
\author{
Younes Boulaguiem, Stéphane Guerrier, Dominique-Laurent Couturier
}
\keyword{internal}
