% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equiv.R
\name{dtost}
\alias{dtost}
\title{The delta-TOST Corrective Procedure for Equivalence Testing}
\usage{
dtost(theta, sigma, nu, alpha, delta)
}
\arguments{
\item{theta}{A \code{numeric} value corresponding to the difference of means.}

\item{sigma}{A \code{numeric} value corresponding to the the standard error.}

\item{nu}{A \code{numeric} value corresponding to the number of degrees of freedom.}

\item{alpha}{A \code{numeric} value specifying the significance level.}

\item{delta}{A \code{numeric} value corresponding to equivalence limit. We assume symmetry, i.e, the equivalence interval corresponds to (-delta,delta)}
}
\value{
A list containing the delta-TOST decision, the corrected confidence interval (CI), estimated mean, estimated standard error,
degrees of freedom, nominal level, equivalence bounds, corrected equivalence bounds, and the method used ("delta-TOST").
}
\description{
This functions is used to compute the delta-TOST, a corrective procedure of the equivalence bounds applied to the Two One-Sided Test (TOST) for equivalence testing in the univariate framework.
}
\examples{
data(skin)

theta_hat = diff(apply(skin,2,mean))
nu = nrow(skin) - 1
sig_hat = sd(apply(skin,1,diff))/sqrt(nu)
res_dtost = dtost(theta = theta_hat, sigma = sig_hat, nu = nu,
              alpha = 0.05, delta = log(1.25))
res_dtost
compare_to_tost(res_dtost)

}
\author{
Younes Boulaguiem, Stéphane Guerrier, Dominique-Laurent Couturier
}
