% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equiv.R
\name{get_alpha_star}
\alias{get_alpha_star}
\title{Get alpha star}
\usage{
get_alpha_star(alpha, sigma_nu, nu, delta, ...)
}
\arguments{
\item{alpha}{A \code{numeric} value specifying the significance level (default = \code{0.05}).}

\item{sigma_nu}{A \code{numeric} value corresponding to the estimated standard error.}

\item{nu}{A \code{numeric} value corresponding to the number of degrees of freedom.}

\item{delta}{A \code{numeric} value corresponding to (bio)equivalence limit. We assume symmetry, i.e, the (bio)equivalence interval corresponds to (-delta,delta)}

\item{...}{Additional parameters.}
}
\value{
The function returns a \code{numeric} value that corresponds to the solution of the optimisation.
}
\description{
Get alpha star
}
\author{
Younes Boulaguiem, Stéphane Guerrier, Dominique-Laurent Couturier
}
\keyword{internal}
