% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equiv_multivariate.R
\name{get_alpha_TOST_MC_mv}
\alias{get_alpha_TOST_MC_mv}
\title{Get alpha star for multivariate TOST}
\usage{
get_alpha_TOST_MC_mv(
  alpha,
  Sigma,
  nu,
  delta,
  theta = NULL,
  B = 10^5,
  tol = .Machine$double.eps^0.5,
  seed = NULL,
  max_iter = 10,
  tolpower = NULL,
  ...
)
}
\arguments{
\item{alpha}{A \code{numeric} value specifying the significance level.}

\item{Sigma}{A \code{numeric} value (univariate) or \code{matrix} (multivariate) corresponding to the estimated variance of estimated \code{theta}.}

\item{nu}{A \code{numeric} value specifying the degrees of freedom. In the multivariate case, it is assumed to be the same across all dimensions.}

\item{delta}{A \code{numeric} value or vector defining the (bio)equivalence margin(s). The procedure assumes symmetry, i.e., the (bio)equivalence region is \eqn{(-\delta, \delta)}.}

\item{theta}{A \code{numeric} value specifying the initial value for the worst-case parameter configuration. If not provided, it's computed via find_sup_x().}

\item{B}{A \code{numeric} value specifying the number of Monte Carlo replication (default: B = \code{10^5}).}

\item{tol}{A \code{numeric} value specifying a tolerance level (default: tol = \code{.Machine$double.eps^0.5}).}

\item{seed}{A \code{character} value representing the use of default seed under \code{NULL}.}

\item{max_iter}{A \code{numeric} value specifying a maximum number of iteration to compute the supremum at which the size is assessed (default: \code{max_iter = 10}).}

\item{tolpower}{A \code{numeric} value specifying the power level to optimize, see Details below for more information.}

\item{...}{Additional parameters.}
}
\value{
The function returns a \code{list} value with the structure:
\itemize{
 \item \code{min}:        A numerical vector that corresponds to the corrected significance level.
 \item \code{theta_sup}:  A numerical vector that corresponds to the evaluated value of supremum.
 \item \code{alphas}:     A numerical vector that corresponds to the significance level used in each iteration.
 \item \code{theta_sups}: A numerical matrix that corresponds to the \code{theta} values evaluated.
 \item \code{powers}:     A numerical vector that corresponds to the estimated power .
 \item \code{iter}:       A numerical variable that corresponds to the number of iterations performed.
 \item \code{err_power}:  A numerical variable that corresponds to the final absolute error between simulated power and nominal significance level \eqn{\alpha}.
}
}
\description{
Get alpha star for multivariate TOST
}
\details{
Tolerance of power allowed between simulated power and nominal alpha. If not specified, it's automatically computed from the 1st and 99th percentiles of a binomial distribution.
}
\author{
Younes Boulaguiem, Luca Insolia, Stéphane Guerrier, Dominique-Laurent Couturier
}
\keyword{internal}
