% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{get_c_of_0}
\alias{get_c_of_0}
\title{TITLE}
\usage{
get_c_of_0(delta, sigma, alpha, B = 1000, tol = 10^(-8), l = 1, optim = "NR")
}
\arguments{
\item{delta}{...}

\item{alpha}{...}

\item{B}{description}

\item{tol}{description}

\item{sig_hat}{...}
}
\description{
What this function does
}
\keyword{internal}
