% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equiv.R
\name{obj_fun_alpha_star}
\alias{obj_fun_alpha_star}
\title{Objective function to optimize in univariate TOST}
\usage{
obj_fun_alpha_star(test, alpha = 0.05, sigma, nu, delta, ...)
}
\arguments{
\item{test}{A \code{numeric} value specifying the significance level to optimize.}

\item{alpha}{A \code{numeric} value specifying the significance level (default: alpha = \code{0.05}).}

\item{sigma}{A \code{numeric} value corresponding to the estimated standard error of estimated \code{theta}.}

\item{nu}{A \code{numeric} value specifying the degrees of freedom.}

\item{delta}{A \code{numeric} value or vector defining the (bio)equivalence margin(s). The procedure assumes symmetry, i.e., the (bio)equivalence region is \eqn{(-\delta, \delta)}.}

\item{...}{Additional parameters.}
}
\value{
The function returns a \code{numeric} value for the objective function.
}
\description{
Objective function to optimize in univariate TOST
}
\author{
Younes Boulaguiem, Stéphane Guerrier, Dominique-Laurent Couturier
}
\keyword{internal}
