% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equiv.R
\name{obj_fun_alpha_star}
\alias{obj_fun_alpha_star}
\title{Objective function to optimise}
\usage{
obj_fun_alpha_star(test, alpha, sigma_nu, nu, delta, ...)
}
\arguments{
\item{test}{A \code{numeric} value specifying the significance level to optimise.}

\item{alpha}{A \code{numeric} value specifying the significance level (default = \code{0.05}).}

\item{sigma_nu}{A \code{numeric} value corresponding to the estimated standard error.}

\item{nu}{A \code{numeric} value corresponding to the number of degrees of freedom.}

\item{delta}{A \code{numeric} value corresponding to (bio)equivalence limit. We assume symmetry, i.e, the (bio)equivalence interval corresponds to (-delta,delta)}

\item{...}{Additional parameters.}
}
\value{
The function returns a \code{numeric} value for the objective function.
}
\description{
Objective function to optimise
}
\author{
Younes Boulaguiem, Stéphane Guerrier, Dominique-Laurent Couturier
}
\keyword{internal}
