% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{obj_fun_alpha_TOST}
\alias{obj_fun_alpha_TOST}
\title{Objective Function of the delta-TOST Corrective Procedure}
\usage{
obj_fun_alpha_TOST(test, alpha, sigma, nu, delta, theta = NULL, ...)
}
\arguments{
\item{test}{The (test) level to be evaluated.}

\item{alpha}{The nominal level for the test.}

\item{sigma}{The considered standard error.}

\item{nu}{The degrees of freedom parameter.}

\item{delta}{The (bio)equivalence bound used for the TOST decision.}

\item{theta}{Parameter value (typically equivalence bound)}

\item{...}{Additional parameters.}
}
\description{
Objective Function of the delta-TOST Corrective Procedure
}
\keyword{internal}
