% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equiv.R
\name{power_TOST}
\alias{power_TOST}
\title{Power function}
\usage{
power_TOST(alpha, theta, sigma_nu, nu, c, ...)
}
\arguments{
\item{alpha}{A \code{numeric} value specifying the significance level (default = \code{0.05}).}

\item{theta}{A \code{numeric} value corresponding to the estimated parameter of interest (such as a difference of means).}

\item{sigma_nu}{A \code{numeric} value corresponding to the the estimated standard error.}

\item{nu}{A \code{numeric} value corresponding to the number of degrees of freedom.}

\item{c}{A \code{numeric} value corresponding to equivalence limit. We assume symmetry, i.e, the equivalence interval corresponds to (-c,c)}

\item{...}{Additional parameters.}
}
\value{
The function returns a \code{numeric} value that corresponds to a probability.
}
\description{
Power function
}
\author{
Younes Boulaguiem, Stéphane Guerrier, etc.
}
