% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equiv_multivariate.R
\name{power_TOST_MC_mv}
\alias{power_TOST_MC_mv}
\title{Power function of univariate or multivariate TOST using Monte Carlo integration}
\usage{
power_TOST_MC_mv(alpha, theta, Sigma, nu, delta, B = 10^5, seed = 10^8)
}
\arguments{
\item{alpha}{A \code{numeric} value specifying the significance level.}

\item{theta}{A \code{numeric} value or vector representing the estimated difference(s) (e.g., between a generic and reference product).}

\item{Sigma}{A \code{numeric} value (univariate) or \code{matrix} (multivariate) corresponding to the estimated variance of estimated \code{theta}.}

\item{nu}{A \code{numeric} value specifying the degrees of freedom. In the multivariate case, it is assumed to be the same across all dimensions.}

\item{delta}{A \code{numeric} value or vector defining the (bio)equivalence margin(s). The procedure assumes symmetry, i.e., the (bio)equivalence region is \eqn{(-\delta, \delta)}.}

\item{B}{A \code{numeric} value specifying the number of Monte Carlo replication (default: B = \code{10^5}).}

\item{seed}{A \code{numeric} value specifying a seed for reproducibility (default: seed = \code{10^8}).}
}
\description{
Power function of univariate or multivariate TOST using Monte Carlo integration
}
\author{
Younes Boulaguiem, Luca Insolia, Stéphane Guerrier, Dominique-Laurent Couturier
}
\keyword{A}
\keyword{The}
\keyword{\code{list}}
\keyword{\code{power_mult}:}
\keyword{\code{power_univ}:}
\keyword{\item}
\keyword{\itemize{}
\keyword{a}
\keyword{corresponds}
\keyword{function}
\keyword{in}
\keyword{internal}
\keyword{multivariate}
\keyword{numerical}
\keyword{probability}
\keyword{returns}
\keyword{setting.}
\keyword{structure:}
\keyword{that}
\keyword{the}
\keyword{to}
\keyword{univariate}
\keyword{value}
\keyword{variable}
\keyword{with}
\keyword{}}
