% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equiv.R
\name{print.tost}
\alias{print.tost}
\title{Print Results of Equivalence Assessment}
\usage{
\method{print}{tost}(x, ticks = 30, rn = 5, ...)
}
\arguments{
\item{x}{An object of class "TOST", either the ouput of the functions `tost`, `atost` or `dtost`.}
}
\description{
Print Results of Equivalence Assessment
}
