% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equiv.R
\name{print.tost}
\alias{print.tost}
\title{Print Results of Equivalence Assessment}
\usage{
\method{print}{tost}(x, ticks = 30, rn = 5, ...)
}
\arguments{
\item{x}{An object of class "TOST", either the ouput of the functions `tost`, `atost` or `dtost`.}

\item{ticks}{Number of ticks to print the confidence interval in the console.}

\item{rn}{Number of digits to consider when printing the results.}

\item{...}{Further arguments to be passed to or from methods.}
}
\description{
Print Results of Equivalence Assessment
}
