% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equiv_quantile.R
\name{qtost_core}
\alias{qtost_core}
\title{Core quantile equivalence testing procedure}
\usage{
qtost_core(
  theta,
  sigma,
  pi_x,
  delta_l,
  delta_u,
  alpha = 0.05,
  corrected_alpha = NULL
)
}
\arguments{
\item{theta}{The calculated test statistic, $\theta$.}

\item{sigma}{The standard error of the test statistic, $\sigma_\theta$.}

\item{pi_x}{A numeric scalar or vector representing the quantile(s) of interest
in the reference population ($X$).}

\item{delta_l}{A numeric scalar or vector for the lower equivalence margin(s).}

\item{delta_u}{A numeric scalar or vector for the upper equivalence margin(s).}

\item{alpha}{The nominal significance level for the test (e.g., 0.05).}

\item{corrected_alpha}{An optional corrected significance level for `alpha-qTOST`.}
}
\value{
An object of class `qtost` or `mqtost` containing the results of the test.
This list includes:
\itemize{
  \item `decision`: A boolean value (`TRUE` for equivalence, `FALSE` otherwise).
  \item `ci`: The confidence interval for the estimated quantile in population Y, $\hat{\pi}_y$.
  \item `pi_y_hat`: The point estimate of the quantile in population Y, calculated as $\Phi(\theta)$.
  \item `theta`: The value of the test statistic $\theta$.
  \item `sigma`: The standard error of $\theta$.
  \item `ci_theta`: The confidence interval for $\theta$.
  \item `alpha`: The significance level used for the test.
  \item `pi_x`: The reference quantile(s).
  \item `delta`: The original equivalence margin.
  \item `eq_region`: The equivalence region defined by `delta_l` and `delta_u`.
  \item `method`: The name of the method used ("qTOST" or "alpha-qTOST").
  \item `setting`: The setting used ("single" or "multiple" quantiles).
}
}
\description{
This is the core the qTOST procedure to obtain CI, etc to test both a single
or multiple quantiles. It takes the test statistic (`theta`), its standard
error (`sigma`), and other parameters to perform the two one-sided tests for
equivalence. This is a low-level function; users should typically use the
main `qtost` wrapper function.
}
\keyword{internal}
