% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate data from the canonical form of the univariate average equivalence problem}
\usage{
simulate_data(mu, sigma, nu, seed = 18)
}
\arguments{
\item{mu}{Numeric value indicating the true mean of the data.}

\item{sigma}{Numeric value indicating the standard deviation of the data.}

\item{nu}{Numeric value indicating the degrees of freedom parameter of the considered setting.}

\item{seed}{Numeric value (optional) used to set the seed for reproducibility of random number generation. Default value is 18.}
}
\value{
A list containing the estimated mean and estimated standard deviation of the simulated data.
}
\description{
Simulate data from the canonical form of the univariate average equivalence problem
}
\examples{
simulate_data(mu = 10, sigma = 2, nu = 10)

}
\keyword{internal}
