% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{size_xTOST}
\alias{size_xTOST}
\title{TITLE}
\usage{
size_xTOST(sig_hat, delta, delta_star, ...)
}
\arguments{
\item{sig_hat}{...}

\item{delta}{...}

\item{delta_star}{...}

\item{...}{description}
}
\description{
What this function does
}
\keyword{internal}
