% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equiv.R
\name{tost}
\alias{tost}
\title{Two One-Sided Test (TOST) for Equivalence Testing}
\usage{
tost(theta, sigma, nu, alpha, delta)
}
\arguments{
\item{theta}{A \code{numeric} value corresponding to the difference of means.}

\item{sigma}{A \code{numeric} value corresponding to the the standard error.}

\item{nu}{A \code{numeric} value corresponding to the number of degrees of freedom.}

\item{alpha}{A \code{numeric} value specifying the significance level.}

\item{delta}{A \code{numeric} value corresponding to equivalence limit. We assume symmetry, i.e, the equivalence interval corresponds to (-delta,delta)}
}
\value{
A list containing the TOST decision, confidence interval (CI), difference of means, standard error,
degrees of freedom, significance level, and the method used ("TOST").
}
\description{
This function performs a Two One-Sided Test (TOST) for equivalence testing.
}
\examples{
data(skin)

theta_hat = diff(apply(skin,2,mean))
nu = nrow(skin) - 1
sig_hat = sd(apply(skin,1,diff))/sqrt(nu)
tost(theta = theta_hat, sigma = sig_hat, nu = nu,
     alpha = 0.05, delta = log(1.25))

}
